requirements.txt
# Telegram Bot
python-telegram-bot==20.7
python-telegram-bot[job-queue]==20.7

# Web Framework for Admin Panel
fastapi==0.109.0
uvicorn==0.27.0
jinja2==3.1.3
python-multipart==0.0.6
itsdangerous==2.1.2

# Database
motor==3.3.2
pymongo==4.6.1

# Configuration
python-dotenv==1.0.0

# HTTP Client for AI API
httpx==0.26.0
aiohttp==3.9.1

# PDF Text Extraction (optional but useful)
PyPDF2==3.0.1

# Authentication & Security
passlib==1.7.4
python-jose[cryptography]==3.3.0
bcrypt==4.1.2

# Utilities
pydantic==2.5.3
pydantic-settings==2.1.0